local_llm_xian_ni/requirement.txt
openai>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
